requirements.txt
configparser
python-dotenv
secure-smtplib
ssl
